+++ src/main/php/TimeZone.php
@@ -60,17 +60,6 @@ class TimeZone
     }
 
     /**
-     * returns internal time zone handle
-     *
-     * @return  \DateTimeZone
-     * @deprecated  since 5.2.0, use handle() instead, will be removed with 6.0.0
-     */
-    public function getHandle()
-    {
-        return $this->handle();
-    }
-
-    /**
      * returns name of time zone
      *
      * @return  string
@@ -81,17 +70,6 @@ class TimeZone
     }
 
     /**
-     * returns name of time zone
-     *
-     * @return  string
-     * @deprecated  since 5.2.0, use name() instead, will be removed with 6.0.0
-     */
-    public function getName()
-    {
-        return $this->name();
-    }
-
-    /**
      * returns offset of the time zone
      *
      * @param   int|string|\DateTime|\stubbles\date\Date  $date  defaults to current date
@@ -99,22 +77,10 @@ class TimeZone
      */
     public function offset($date = null)
     {
-        $offset  = $this->getOffsetInSeconds($date);
+        $offset  = $this->offsetInSeconds($date);
         $hours   = intval(abs($offset) / 3600);
         $minutes = (abs($offset)- ($hours * 3600)) / 60;
         return sprintf('%s%02d%02d', ($offset < 0 ? '-' : '+'), $hours, $minutes);
-    }
-
-    /**
-     * returns offset of the time zone
-     *
-     * @param   int|string|\DateTime|\stubbles\date\Date  $date  defaults to current date
-     * @return  string
-     * @deprecated  since 5.2.0, use offset() instead, will be removed with 6.0.0
-     */
-    public function getOffset($date = null)
-    {
-        return $this->offset($date);
     }
 
     /**
@@ -134,22 +100,6 @@ class TimeZone
         }
 
         return $this->timeZone->getOffset(Date::castFrom($date)->handle());
-    }
-
-    /**
-     * returns offset to given date in seconds
-     *
-     * Because a timezone may have different offsets when its in DST or non-DST
-     * mode, a date object must be given which is used to determine whether DST
-     * or non-DST offset should be returned.
-     *
-     * @param   int|string|\DateTime|\stubbles\date\Date  $date  defaults to current date
-     * @return  int
-     * @deprecated  since 5.2.0, use offsetInSeconds() instead, will be removed with 6.0.0
-     */
-    public function getOffsetInSeconds($date = null)
-    {
-        return $this->offsetInSeconds($date);
     }
 
     /**