+++ src/main/php/span/Datespan.php
@@ -23,14 +23,6 @@ interface Datespan
     public function start();
 
     /**
-     * returns the start date
-     *
-     * @return  \stubbles\date\Date
-     * @deprecated  since 5.2.0, use start() instead, will be removed with 6.0.0
-     */
-    public function getStart();
-
-    /**
      * checks whether datespan starts before a given date
      *
      * @param   int|string|\DateTime|\stubbles\date\Date  $date
@@ -54,14 +46,6 @@ interface Datespan
      * @return  \stubbles\date\Date
      */
     public function end();
-
-    /**
-     * returns the end date
-     *
-     * @return  \stubbles\date\Date
-     * @deprecated  since 5.2.0, use end() instead, will be removed with 6.0.0
-     */
-    public function getEnd();
 
     /**
      * checks whether datespan ends before a given date
@@ -108,13 +92,6 @@ interface Datespan
      */
     public function amountOfDays();
 
-    /**
-     * returns amount of days in this datespan
-     *
-     * @return  int
-     * @deprecated  since 5.2.0, use amountOfDays() instead, will be removed with 6.0.0
-     */
-    public function getAmountOfDays();
 
     /**
      * returns list of days
@@ -122,13 +99,6 @@ interface Datespan
      * @return  \stubbles\date\span\Day[]
      */
     public function days();
-    /**
-     * returns list of days
-     *
-     * @return  \stubbles\date\span\Day[]
-     * @deprecated  since 5.2.0, use days() instead, will be removed with 6.0.0
-     */
-    public function getDays();
 
     /**
      * checks whether the span is in the future compared to current date